Guess leftmost strip by maximum relative error
The "rightmost strip will have the worst match to the right" heuristic
fails when the image is highly variant in that region -- a 'bad' match
may in fact be exactly inline with what the image "looks like" in that
area.

This patch first determines the errors on either side and computes the
relative error our match produces.  Guessing that the rightmost is the
highest with this new relative works in all my tests cases and the Tokyo
panorama (fixes the black/white skyscraper), but fails on the Mona Lisa
test image.