Guess leftmost as the strip nobody wants to their right
Since we ask for 'best right matches', if a single strip is never
matched as the 'best right match' for any other strip, let's assume it's
the leftmost.  If we can't find any, or if multiple strips were never
matched, we fall back to the previous 'highest average error' algorithm.