Implement strip-width guessing
Finally got around to implementing this.  The idea is to figure out the
mean and standard deviation for the distances between adjacent pixel
columns, and then figure out the most common width between distances
which are more than 2 standard deviations from the mean.

It works quite well, only failing on one test image I have which is
pathological for this particular algorithm: it's an overhead view of a
city block with lots of veritical lines with large colour changes
between them as buildings and streets start and end.